Changes to graphics updating and added maze borders
Changed how often graphics update to only update every time the player moves. This reduces resource consumption when the player is not moving. This also acts as a temporary and only partial fix to the flickering graphics, as the flickering still occurs occasionally. The colour of the player has also been changed to red to allow for constrast with the walls.